refactor: split out spawnServer and setupAdmin
