+++ main.js
@@ -2,11 +2,8 @@ require('dotenv').config();
 const fs = require('fs-extra');
 const assert = require("assert");
 const { Readable, Writable } = require("stream");
-const readline = require("readline");
 let { spawn, exec } = require("child_process");
 const pidusage = require("pidusage");
-const express = require("express");
-const crypto = require("crypto");
 
 const {
     UNZIPPED_SERVERS_CONTAINER,
@@ -14,9 +11,12 @@ const {
     BACKUP_TYPES,
     MS_IN_MIN,
     MS_IN_SEC,
+    SAVE_QUERY_FREQUENCY,
     BUCKET_LOCK_FILE_NAME,
     formatBytes,
-    platform
+    platform,
+    newlog,
+    sec2time
 } = require("./utils.js");
 
 if (platform === "win32") {
@@ -39,11 +39,7 @@ const {
     deleteLockFileIfExists,
 } = require("./backup.js");
 
-const rl = readline.createInterface({
-    input: process.stdin,
-    output: process.stdout,
-    terminal: false
-});
+const { setupAdmin } = require("./admin.js");
 
 fs.ensureDirSync(UNZIPPED_SERVERS_CONTAINER);
 
@@ -54,239 +50,103 @@ assert(backupConfig, `Could not find field 'backup' at root of config`);
 
 const backupFrequencyMS = backupConfig["backup-frequency-minutes"] * MS_IN_MIN;
 const minBackupFrequencyMinutes = 10;
-assert(
-    backupFrequencyMS > MS_IN_MIN * minBackupFrequencyMinutes,
-    `Expected backup['backup-frequency-min'] to be greater than ${minBackupFrequencyMinutes}`
-);
+
+assert(backupFrequencyMS > MS_IN_MIN * minBackupFrequencyMinutes,
+    `Expected backup['backup-frequency-min'] to be greater than ${minBackupFrequencyMinutes}`);
+
+assert(platform === "linux" || platform === 'win32',
+    'Unsupported platform - must be Windows 10 or Ubuntu 18+ based');
+
+
 let isCurrentlyBackingUp = false;
 let hasSentStopCommand = false;
-const SAVE_QUERY_FREQUENCY = MS_IN_SEC * 5;
-const UI_COMMAND_DELAY = MS_IN_SEC * 1;
-
 let currentBackupType = null;
 
-function sec2time(timeInSeconds) {
-    var pad = function(num, size) {
-            return ("000" + num).slice(size * -1);
-        },
-        time = parseFloat(timeInSeconds).toFixed(3),
-        hours = Math.floor(time / 3600),
-        minutes = Math.floor(time / 60) % 60,
-        seconds = Math.floor(time - minutes * 60),
-        milliseconds = time.slice(-3);
-    return (
-        pad(hours, 2) +
-        ":" +
-        pad(minutes, 2) +
-        ":" +
-        pad(seconds, 2) +
-        "." +
-        pad(milliseconds, 3)
-    );
-}
+console.log = newlog;
 
-const MAX_STORED_LINES = 20;
-const consoleLogBuffer = [];
-const originalConsoleLog = console.log;
+const spawnServer = (args) => {
+    const { saveQueryInterval, saveHoldInterval } = args;
 
-console.log = text => {
-    originalConsoleLog(text);
-    consoleLogBuffer.push((text || "").toString().replace(/\n/, "<br>"));
-    if (consoleLogBuffer.length > MAX_STORED_LINES) {
-        consoleLogBuffer.shift(); // delete first item.
-    }
-};
+    const bs = spawn("./bedrock_server", [], {
+        stdio: ["pipe", "pipe", "pipe", "ipc"],
+        cwd: UNZIPPED_SERVER_FOLDER_PATH
+    });
 
-let salt = crypto
-    .randomBytes(32)
-    .toString("hex")
-    .toUpperCase();
+    bs.stderr.on("data", error => {
+        console.error(`SERVER STDERROR: ${error}`);
+    });
 
-const uiConfig = config.ui;
-if ((uiConfig || {}).enabled) {
-    console.log("Starting express server");
-    
-    function clientHashIsValid(clientHashWithSalt) {
-        return (
-            (clientHashWithSalt || "").toUpperCase() ===
-            crypto
-            .createHash("sha256")
-            .update(
-                uiConfig["admin-code-sha256-hash"].toUpperCase() +
-                salt.toUpperCase()
-            )
-            .digest("hex")
-            .toUpperCase()
-        );
-    }
+    bs.on("error", data => {
+        console.log(`SERVER ERROR: ${data}`);
+    });
 
-    const expressApp = express();
-    const router = express.Router();
-    expressApp.use(express.json());
-
-    router.get("/terminal-out", (req, res) => {
-        res.send(
-            [`${MAX_STORED_LINES} latest lines of terminal output:`]
-            .concat(consoleLogBuffer)
-            .join("<br>")
+    bs.on("close", code => {
+        console.log(
+            `Minecraft server child process exited with code ${code}`
         );
     });
 
-    router.get("/resource-usage", (req, res) => {
-        if (bs != null) {
-            pidusage(bs.pid, (err, stats) => {
-                let result = {};
-                if (stats != null) {
-                    result = {
-                        cpu: stats.cpu,
-                        elapsed: stats.elapsed,
-                        memory: stats.memory
-                    };
-                }
-                res.send(result);
-            });
+    bs.stdout?.on("data", async data => {
+        if (
+            /^(A previous save has not been completed\.|Saving\.\.\.|Changes to the level are resumed\.)/i.test(
+                data
+            )
+        ) {
+            // do nothing
+        } else if (
+            /^(Data saved\. Files are now ready to be copied\.)/i.test(
+                data
+            )
+        ) {
+            isCurrentlyBackingUp = true;
+
+            const backupStartTime = Math.floor(new Date() / 1000);
+            const backupType = currentBackupType;
+            console.log(
+                `Files ready for backup! Creating backup of server state at ${new Date(
+                    backupStartTime * MS_IN_SEC
+                ).toLocaleString()} with type ${backupType}...`
+            );
+
+            const dataSplit = data
+                .toString()
+                .split(
+                    "Data saved. Files are now ready to be copied."
+                );
+            const backupFileListString = dataSplit[
+                dataSplit.length - 1
+            ].replace(/(\n|\r|\\n|\\r)/g, "");
+            await createBackup(
+                backupFileListString,
+                backupStartTime,
+                backupType
+            );
+            isCurrentlyBackingUp = false;
+            bs.stdin.write("save resume\r\n");
+            // stop here, since the backup before stop has completed;
+            if (hasSentStopCommand) {
+                clearInterval(saveQueryInterval);
+                clearInterval(saveHoldInterval);
+                bs.stdin.write("stop\r\n");
+                setTimeout(async () => {
+                    await deleteLockFileIfExists();
+                    process.exit(0);
+                }, MS_IN_SEC);
+            }
         } else {
-            res.send({});
+            console.log(`${data.toString().replace(/\n$/, "")}`);
         }
     });
+    return bs;
+};
 
-    router.get("/salt", (req, res) => {
-        // refresh salt anytime anyone asks for it
-        salt = crypto
-            .randomBytes(32)
-            .toString("hex")
-            .toUpperCase();
-        res.send(salt);
-    });
-
-    router.get("/is-auth-valid", (req, res) => {
-        res.send(clientHashIsValid(req.header("Authorization")));
-    });
-
-    router.post("/stop", (req, res) => {
-        const {
-            body
-        } = req;
-        const {
-            adminCodeHash
-        } = {
-            body
-        };
-        setTimeout(() => {
-            if (clientHashIsValid(req.header("Authorization"))) {
-                rl.write("stop\n");
-                res.sendStatus(200);
-            } else {
-                console.log("Rejected unauthorized request to stop server");
-                res.sendStatus(401);
-            }
-        }, UI_COMMAND_DELAY);
-    });
-
-    router.post("/trigger-manual-backup", (req, res) => {
-        const {
-            body
-        } = req;
-        const {
-            adminCodeHash
-        } = {
-            body
-        };
-        setTimeout(() => {
-            if (clientHashIsValid(req.header("Authorization"))) {
-                rl.write("backup\n");
-                res.sendStatus(200);
-            } else {
-                console.log(
-                    "Rejected unauthorized request to trigger manual backup"
-                );
-                res.sendStatus(401);
-            }
-        }, UI_COMMAND_DELAY);
-    });
-
-    router.post("/trigger-print-resource-usage", (req, res) => {
-        const {
-            body
-        } = req;
-        setTimeout(() => {
-            if (clientHashIsValid(req.header("Authorization"))) {
-                rl.write("resource-usage\n");
-                res.sendStatus(200);
-            } else {
-                console.log(
-                    "Rejected unauthorized request to print resource usage"
-                );
-                res.sendStatus(401);
-            }
-        }, UI_COMMAND_DELAY);
-    });
-
-    router.post("/trigger-print-player-list", (req, res) => {
-        const {
-            body
-        } = req;
-        setTimeout(() => {
-            if (clientHashIsValid(req.header("Authorization"))) {
-                rl.write("list\n");
-                res.sendStatus(200);
-            } else {
-                console.log(
-                    "Rejected unauthorized request to print resource usage"
-                );
-                res.sendStatus(401);
-            }
-        }, UI_COMMAND_DELAY);
-    });
-
-    router.post("/trigger-restore-backup", async (req, res) => {
-        const {
-            body
-        } = req;
-        const {
-            adminCodeHash
-        } = {
-            body
-        };
-        setTimeout(async () => {
-            if (clientHashIsValid(req.header("Authorization"))) {
-                const backup = body.backup;
-                const backups = await getBackupList();
-                if (backups.includes(backup)) {
-                    // do this check to avoid weird injection errors
-                    rl.write(`force-restore ${body.backup}\n`);
-                } else {
-                    console.log(`Backup ${body.backup} not found`);
-                }
-                res.sendStatus(200);
-            } else {
-                console.log(
-                    `Rejected unauthorized request to restore backup ${body.backup}`
-                );
-                res.sendStatus(401);
-            }
-        }, UI_COMMAND_DELAY);
-    });
-
-    router.get("/backup-size-list", async (req, res) => {
-        const backups = await getBackupSizeList();
-        res.send(backups);
-    });
-    
-    expressApp.use("/", router);
-    expressApp.use(express.static("static"));
-    expressApp.listen(uiConfig.port);
-
-    console.log(`Running UI for server on port ${uiConfig.port}`);
-} else {
-    console.log("Running server without UI");
-}
-
-let bs = null;
 
 createServerProperties().then(async () => {
     await createBackupBucketIfNotExists();
     const doesLockFileExist = await doesLockFileExistOrS3Disabled();
+    let saveQueryInterval = null;
+    let saveHoldInterval = null;
+
     if (doesLockFileExist) {
         console.error(`Exiting server; a lock file, ${BUCKET_LOCK_FILE_NAME} exists within the S3 backup bucket.`
             + ' This usually means another server is connected to the backup bucket.'
@@ -300,88 +160,13 @@ createServerProperties().then(async () => {
 
     console.log(`Starting Minecraft Bedrock server in ${process.env.ENVIRONMENT} mode...`);
     console.log(`Path: ${UNZIPPED_SERVER_FOLDER_PATH}/bedrock_server`)
-    const spawnServer = () => {
-        bs = spawn("./bedrock_server", [], {
-            stdio: ["pipe", "pipe", "pipe", "ipc"],
-            cwd: UNZIPPED_SERVER_FOLDER_PATH
-        });
-
-        bs.stderr.on("data", error => {
-            console.error(`SERVER STDERROR: ${error}`);
-        });
-
-        bs.on("error", data => {
-            console.log(`SERVER ERROR: ${data}`);
-        });
-
-        bs.on("close", code => {
-            console.log(
-                `Minecraft server child process exited with code ${code}`
-            );
-        });
-
-        bs.stdout.on("data", async data => {
-            if (
-                /^(A previous save has not been completed\.|Saving\.\.\.|Changes to the level are resumed\.)/i.test(
-                    data
-                )
-            ) {
-                // do nothing
-            } else if (
-                /^(Data saved\. Files are now ready to be copied\.)/i.test(
-                    data
-                )
-            ) {
-                isCurrentlyBackingUp = true;
-
-                const backupStartTime = Math.floor(new Date() / 1000);
-                const backupType = currentBackupType;
-                console.log(
-                    `Files ready for backup! Creating backup of server state at ${new Date(
-                        backupStartTime * MS_IN_SEC
-                    ).toLocaleString()} with type ${backupType}...`
-                );
-
-                const dataSplit = data
-                    .toString()
-                    .split(
-                        "Data saved. Files are now ready to be copied."
-                    );
-                backupFileListString = dataSplit[
-                    dataSplit.length - 1
-                ].replace(/(\n|\r|\\n|\\r)/g, "");
-                await createBackup(
-                    backupFileListString,
-                    backupStartTime,
-                    backupType
-                );
-                isCurrentlyBackingUp = false;
-                bs.stdin.write("save resume\r\n");
-                // stop here, since the backup before stop has completed;
-                if (hasSentStopCommand) {
-                    clearInterval(saveQueryInterval);
-                    clearInterval(saveHoldInterval);
-                    bs.stdin.write("stop\r\n");
-                    setTimeout(async () => {
-                        await deleteLockFileIfExists();
-                        process.exit(0);
-                    }, MS_IN_SEC);
-                }
-            } else {
-                console.log(`${data.toString().replace(/\n$/, "")}`);
-            }
-        });
-    };
-
-    if (platform === "linux" || platform === 'win32') {
-        spawnServer();
-    } else {
-        throw 'Unsupported platform - must be Windows 10 or Ubuntu 18+ based';
-    }
-
+    
+    const bs = spawnServer({ saveQueryInterval, saveHoldInterval });
+    let { rl } = setupAdmin(bs);
+    
     let lastQueryWasSaveSucccessful = false;
 
-    const saveQueryInterval = setInterval(() => {
+    saveQueryInterval = setInterval(() => {
         if (!isCurrentlyBackingUp && bs) {
             bs.stdin.write("save query\r\n");
         }
@@ -396,7 +181,7 @@ createServerProperties().then(async () => {
         }
     };
 
-    const saveHoldInterval = process.env.SCHEDULE_BACKUPS === 'true' ? setInterval(
+    saveHoldInterval = process.env.SCHEDULE_BACKUPS === 'true' ? setInterval(
         triggerBackup,
         backupFrequencyMS,
         BACKUP_TYPES.SCHEDULED
