+++ package.json
@@ -23,6 +23,7 @@
     "aws-sdk": "^2.656.0",
     "cross-spawn": "^7.0.2",
     "dotenv": "^16.0.3",
+    "eslint": "^8.28.0",
     "express": "^4.16.3",
     "fs-extra": "^9.0.0",
     "lodash": "^4.17.21",
